refactor: extract find in files
